requirements.txt
PyContracts>=1.8.12,<2
PyYAML>=5.2,<6
tabulate>=0.8.6,<0.9
